LayoutTests/platform/nix/fast/exclusions/shape-outside-floats/shape-outside-floats-ellipse-margin-right-expected.txt

PASS elementRect('s1').top is 0
PASS elementRect('s1').right is 127
PASS elementRect('s2').top is 20
PASS elementRect('s2').right is 102
PASS elementRect('s3').top is 40
FAIL elementRect('s3').right should be 86. Was 87.
The right edges of the three black squares should follow the outer ellipse boundary and each square should appear on a subsequent line.

X
X
X
